refactor: move to memberRepository
