+++ src/main/java/pretzel/dreamketcherbe/domain/webtoon/repository/WebtoonRepository.java
@@ -1,7 +1,6 @@
 package pretzel.dreamketcherbe.domain.webtoon.repository;
 
 import io.lettuce.core.dynamic.annotation.Param;
-import java.util.List;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -25,18 +24,6 @@ public interface WebtoonRepository extends JpaRepository<Webtoon, Long>, Webtoon
     Page<Webtoon> findByTitle(@Param("keyword") String keyword, Pageable pageable);
 
     Page<Webtoon> findAllByOrderByCreatedAtDesc(Pageable pageable);
-
-    @Query("""
-            SELECT DISTINCT w.member
-            FROM Webtoon w
-            WHERE (LOWER(w.member.nickname) LIKE %:keyword%
-            OR LOWER(REPLACE(w.member.nickname,' ','')) LIKE %:keyword%)
-            AND w.isDeleted = false
-        """)
-    List<Member> findDistinctMembersByNickname(
-        @Param("keyword") String keyword,
-        Pageable pageable
-    );
 
     // 특정 작가의 첫 작품 (등록일 오름차순)
     Optional<Webtoon> findFirstByMemberAndIsDeletedFalseOrderByCreatedAtAsc(Member member);
